+++ Sources/StreamChatUI/ChatMessageList/Reactions/ChatReactionPickerReactionsView.swift
@@ -1,0 +1,9 @@
+//
+// Copyright © 2021 Stream.io Inc. All rights reserved.
+//
+
+open class ChatReactionPickerReactionsView: ChatMessageReactionsView {
+    override public var reactionItemView: ChatMessageReactionItemView.Type {
+        components.reactionPickerReactionItemView
+    }
+}