+++ Sources/StreamChat/Models/Attachments/ChatMessageFileAttachment.swift
@@ -33,6 +33,16 @@ public struct FileAttachmentPayload: AttachmentPayload {
             .flatMap { try? JSONEncoder.stream.encode($0) }
             .flatMap { try? JSONDecoder.stream.decode(T.self, from: $0) }
     }
+    
+    /// Creates `FileAttachmentPayload` instance.
+    ///
+    /// Use this initializer if the attachment is already uploaded and you have the remote URLs.
+    public init(title: String?, assetRemoteURL: URL, file: AttachmentFile, extraData: [String: RawJSON]?) {
+        self.title = title
+        assetURL = assetRemoteURL
+        self.file = file
+        self.extraData = extraData
+    }
 }
 
 extension FileAttachmentPayload: Hashable {}
@@ -58,7 +68,7 @@ extension FileAttachmentPayload: Decodable {
         
         self.init(
             title: try container.decodeIfPresent(String.self, forKey: .title),
-            assetURL: try container.decode(URL.self, forKey: .assetURL),
+            assetRemoteURL: try container.decode(URL.self, forKey: .assetURL),
             file: try AttachmentFile(from: decoder),
             extraData: try Self.decodeExtraData(from: decoder)
         )